backend/requirements.txt
# Core
Django>=4.2,<5.0
graphene-django>=3.1
django-cors-headers>=4.3
django-filter>=23.5
gunicorn>=21.2
whitenoise>=6.6

# Environment & config
django-environ>=0.11
python-dotenv>=1.0

# Database (PostgreSQL)
psycopg2-binary>=2.9

# Async task queue
celery>=5.3
redis>=5.0

# Authentication (JWT)
django-graphql-jwt>=0.4

# Google OAuth
google-auth>=2.25

# HTTP probing & system metrics
requests>=2.31
psutil>=5.9

# Payments
razorpay>=1.4

# Real analysis tools
radon>=6.0
bandit>=1.7
safety>=2.3

# Error monitoring
sentry-sdk[django]>=1.40

# Structured logging
python-json-logger>=2.0

# ML/LLM & Transformer Models
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
langchain>=0.1.20
langchain-community>=0.0.15
langchain-core>=0.1.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
numpy>=1.24.0

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3

psycopg2-binary>=2.9
celery>=5.2
redis>=4.5
django-graphql-jwt>=0.3.0
django-environ>=0.9.0
sentry-sdk>=1.5.0
django-structlog>=3.0
bandit>=1.7
safety>=1.10.0
radon>=1.0.0
pytest>=7.0
pytest-django>=4.0
pytest-cov>=3.0
factory-boy>=3.2
coverage>=6.0